+++ gatr/__init__.py
@@ -1,8 +1,8 @@
-# Copyright (c) 2023 Qualcomm Technologies, Inc.
+# Copyright (c) 2024 Qualcomm Technologies, Inc.
 # All rights reserved.
 from .layers.attention.config import SelfAttentionConfig
 from .layers.mlp.config import MLPConfig
 from .nets.axial_gatr import AxialGATr
 from .nets.gatr import GATr
 
-__version__ = "1.3.0"
+__version__ = "1.4.0"