+++ gatr/interface/__init__.py
@@ -1,4 +1,4 @@
-# Copyright (c) 2023 Qualcomm Technologies, Inc.
+# Copyright (c) 2024 Qualcomm Technologies, Inc.
 # All rights reserved.
 from .object import (
     embed_3d_object,
@@ -9,6 +9,7 @@ from .object import (
 from .plane import embed_oriented_plane, extract_oriented_plane
 from .point import embed_point, extract_point, extract_point_embedding_reg
 from .pseudoscalar import embed_pseudoscalar, extract_pseudoscalar
+from .ray import embed_pluecker_ray, extract_pluecker_ray
 from .reflection import embed_reflection, extract_reflection
 from .rotation import embed_rotation, extract_rotation
 from .scalar import embed_scalar, extract_scalar