+++ gatr/utils/warning.py
@@ -1,0 +1,6 @@
+# Copyright (c) 2024 Qualcomm Technologies, Inc.
+# All rights reserved.
+
+
+class GATrDeprecationWarning(UserWarning):
+    """Warning about a deprecation in the GATr library."""